SA-2209 - Move component back to its original location
