+++ src/pages/reportBuilder/components/index.js
@@ -3,8 +3,9 @@ export { default as ActiveMetrics } from './ActiveMetrics';
 export { default as Charts } from './Charts';
 export { default as CompareByForm } from './CompareByForm';
 export { default as FiltersForm } from './FiltersForm';
+export { default as GroupByTable } from './GroupByTable/GroupByTable';
+export { default as CompareByGroupByTable } from './GroupByTable/CompareByGroupByTable';
+export { IndustryBenchmarkModal } from './IndustryBenchmarkModal';
 export { default as MetricsForm } from './MetricsForm';
 export { default as ReportOptions } from './ReportOptions';
 export { default as SaveReportModal } from './SavedReportsSection/SaveReportModal';
-export { default as GroupByTable } from './GroupByTable/GroupByTable';
-export { default as CompareByGroupByTable } from './GroupByTable/CompareByGroupByTable';
